Run integration tests (#276)
* Initial attempt at running integration tests

* Fix workflow syntax error

* Another syntax error

* Can't use an env value to define another env value :-(

* Use powershell core for secrets handling

* Correct indentation

* Get the namespace right

* Disable fail-fast for matrix

* Only set secrets env var where it is needed

* Syntax error

* Install hostedwebcore

* Enable all namespaces

* Built profiler is now included in the homedirs

* Ignore flickering test until next month

* Do a pull after comitting test results changes in order to mitigate potential race condition

* Remove parent WCF namespaces to avoid running the same tests multiple times

* Namespace for data transmission tests

* Namespaces for errors and agent metrics (supportability and CLR)
