+++ tests/Agent/IntegrationTests/IntegrationTests/AgentMetrics/MemoryMvc.cs
@@ -10,14 +10,14 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests.AgentFeatures
+namespace NewRelic.Agent.IntegrationTests.AgentMetrics
 {
     [NetFrameworkTest]
-    public class CpuMvc : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
+    public class MemoryMvc : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
     {
         private readonly RemoteServiceFixtures.BasicMvcApplicationTestFixture _fixture;
 
-        public CpuMvc(RemoteServiceFixtures.BasicMvcApplicationTestFixture fixture, ITestOutputHelper testLogger)
+        public MemoryMvc(RemoteServiceFixtures.BasicMvcApplicationTestFixture fixture, ITestOutputHelper testLogger)
         {
             _fixture = fixture;
             _fixture.TestLogger = testLogger;
@@ -33,7 +33,7 @@ namespace NewRelic.Agent.IntegrationTests.AgentFeatures
                     var startTime = DateTime.Now;
                     while (DateTime.Now <= startTime.AddSeconds(60))
                     {
-                        if (_fixture.AgentLog.GetMetrics().Any(metric => metric.MetricSpec.Name == "CPU/User Time"))
+                        if (_fixture.AgentLog.GetMetrics().Any(metric => metric.MetricSpec.Name == "Memory/Physical"))
                             break;
                         Thread.Sleep(TimeSpan.FromSeconds(5));
                     }
@@ -47,8 +47,7 @@ namespace NewRelic.Agent.IntegrationTests.AgentFeatures
         {
             var expectedMetrics = new List<Assertions.ExpectedMetric>
             {
-                new Assertions.ExpectedMetric {metricName = @"CPU/User Time"},
-                new Assertions.ExpectedMetric {metricName = @"CPU/User/Utilization"}
+                new Assertions.ExpectedMetric {metricName = @"Memory/Physical"}
             };
 
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
