+++ tests/Agent/IntegrationTests/IntegrationTests/Errors/ErrorTraceWebApi.cs
@@ -11,7 +11,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests.AgentFeatures
+namespace NewRelic.Agent.IntegrationTests.Errors
 {
     [NetFrameworkTest]
     public class ErrorTraceWebApi : IClassFixture<RemoteServiceFixtures.OwinWebApiFixture>
