+++ tests/Agent/IntegrationTests/IntegrationTests/Errors/ErrorTraceWebService.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests.AgentFeatures
+namespace NewRelic.Agent.IntegrationTests.Errors
 {
     [NetFrameworkTest]
     public class ErrorTraceWebService : IClassFixture<RemoteServiceFixtures.BasicWebService>
